chore: update dependencies and versions across the project
- Bump workspace version to 2.5.0
- Update polars and related dependencies to version 0.55.1
- Upgrade lz4_flex to version 0.14.0
- Update pyo3 and pyo3-polars to versions 0.29.1 and 0.28.0 respectively
- Adjust timestamp calculation in connector.rs for better precision

+++ py-kola/src/connector.rs
@@ -93,7 +93,7 @@ fn cast_k_to_py(py: Python, k: K) -> PyResult<Py<PyAny>> {
             let mut days = k.num_days_from_ce() as i64 - 719163;
             days = min(days, 2932532);
             days = max(days, -719162);
-            let date = PyDate::from_timestamp(py, 86400 * days)?;
+            let date = PyDate::from_timestamp(py, 86400.0 * days as f64)?;
             date.into_py_any(py)
         }
         K::Time(k) => {
